api/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-nested-routers==0.93.5
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
